+++ src/main/java/com/huayuan/web/dto/CreditLimitDto.java
@@ -4,22 +4,22 @@ package com.huayuan.web.dto;
  * Created by dell on 14-4-14.
  */
 public class CreditLimitDto {
-    private Integer creditLimit;
+    private Double creditLimit;
     private String rankOfLimit;
 
     public CreditLimitDto() {
     }
 
-    public CreditLimitDto(Integer creditLimit, String rankOfLimit) {
+    public CreditLimitDto(Double creditLimit, String rankOfLimit) {
         this.creditLimit = creditLimit;
         this.rankOfLimit = rankOfLimit;
     }
 
-    public Integer getCreditLimit() {
+    public Double getCreditLimit() {
         return creditLimit;
     }
 
-    public void setCreditLimit(Integer creditLimit) {
+    public void setCreditLimit(Double creditLimit) {
         this.creditLimit = creditLimit;
     }
